fix: redirects for newly installed or deleted packages
