+++ internal/web/templates/layout/base.html
@@ -6,7 +6,7 @@
     <link type="text/css" rel="stylesheet" href="/static/bundle/index.min.css?v={{ .CacheBustingString }}" />
     <script src="/static/bundle/index.min.js?v={{ .CacheBustingString }}"></script>
     <script type="text/javascript">
-      htmx.logAll();
+      // htmx.logAll();
     </script>
   </head>
